+++ lib/Controller/SettingsController.php
@@ -155,7 +155,8 @@ class SettingsController extends Controller {
             "tagsEnabled" => \OC::$server->getAppManager()->isEnabledForUser("systemtags"),
             "reviewDisplay" => $this->config->getCustomizationReviewDisplay(),
             "theme" => $this->config->getCustomizationTheme(),
-            "templates" => $this->getGlobalTemplates()
+            "templates" => $this->getGlobalTemplates(),
+            "unknownAuthor" => $this->config->getUnknownAuthor()
         ];
         return new TemplateResponse($this->appName, "settings", $data, "blank");
     }
@@ -235,6 +236,7 @@ class SettingsController extends Controller {
      * @param bool $help - display help
      * @param bool $toolbarNoTabs - display toolbar tab
      * @param string $reviewDisplay - review viewing mode
+     * @param string $unknownAuthor - display unknown author
      *
      * @return array
      */
@@ -254,7 +256,8 @@ class SettingsController extends Controller {
         $help,
         $toolbarNoTabs,
         $reviewDisplay,
-        $theme
+        $theme,
+        $unknownAuthor
     ) {
 
         $this->config->setDefaultFormats($defFormats);
@@ -273,6 +276,7 @@ class SettingsController extends Controller {
         $this->config->setCustomizationToolbarNoTabs($toolbarNoTabs);
         $this->config->setCustomizationReviewDisplay($reviewDisplay);
         $this->config->setCustomizationTheme($theme);
+        $this->config->setUnknownAuthor($unknownAuthor);
 
         return [
         ];