+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## Added
+- setting for setup unknown author display name
+
+## Changed
+- fix address when opening editor
+
 ## 9.5.0
 ## Added
 - default empty templates